+++ lib/core/facets/kernel/tap.rb
@@ -1,3 +1,3 @@
-# Facets no longer overrides Kernel#tap. Use Kernel#tee instead
+# Facets no longer overrides Kernel#tap. Use Kernel#functor instead
 # for the block-less Functor form.
-require 'facets/kernel/tee'
+require 'facets/kernel/functor'